+++ NEWS.md
@@ -7,7 +7,21 @@
       some columns must be numeric; check for numeric? Not yet implemented
 * [ ] Allow for custom user functions when calling annex_stats?
 
-# Annex 0.1
+# annex 0.2-0
+
+Under development since January 2, 2023; intended to be the next release
+after the initial testing phase December 2022 to February 2023.
+
+* Renaming 'Pollutant' consistently to 'Variable' (issue #2)
+* Renaming sheet 'Meta-Period' to 'Meta Season' (issue #2)
+* Automatically filling 'Measurement Location' in META-Room (issue #3)
+* Automatically filling 'Variable Name' in META-Variable (issue #4)
+* Adjusting some (internal) function names (Pollutants -> Variables; Period -> Season)
+    given the adjustments updates listed above.
+
+
+
+# annex 0.1
 
 Summary of major fixes and features during the development of the
 package in its `0.1-X` state.